temp: encode / decode app player

+++ Source/Debug/AudioAppPlayer.h
@@ -24,7 +24,7 @@ public:
     void onAudioBlockProcessedEvent(const AudioBlockProcessedEvent& event) override;
 
 private:
-    std::vector<int16_t> audioBlock; // Contient les échantillons audio
+    std::vector<float> audioBlock; // Contient les échantillons audio
     std::vector<int8_t> opusData; // Contient les données audio encodées
     int currentNumSamples = 0; // Nombre d'échantillons dans le bloc
     size_t currentSampleIndex = 0; // Position actuelle dans le bloc
