+++ Source/Rtc/WebRTCAudioService.h
@@ -30,7 +30,7 @@ private:
     uint32_t ssrc = 12345; // Identifiant SSRC unique
     // Audio Thread
     std::thread audioThread;
-    std::queue<std::vector<int16_t>> audioQueue;
+    std::queue<std::vector<float>> audioQueue;
     std::mutex queueMutex;
     std::condition_variable queueCondition;
     bool audioThreadRunning = false;
